[TensorExpr] TensorExprKernel: don't do any compilation or lowering in run().
The input JIT graph has all the information we need to perform the
entire compilation at the construction time. We don't need to postpone
any steps until the execution time. Also, from the graph we always know
what device we will be executing on and thus we don't need to have a
CodeGen cache in TensorExprKernel - we always have one and only one
CodeGen.

ghstack-source-id: 90856fa5e3d460b33bf6668d63b29a4e2a25b212
Pull Request resolved: https://github.com/pytorch/pytorch/pull/37948